Changed Routes, Added Auth
