+++ src/lib/files/typecheck.ts
