+++ src/routes/modules/+layout.svelte
@@ -1,0 +1,7 @@
+<script>
+  import "../../app.css";
+</script>
+
+<main>
+  <slot />
+</main>
