+++ src/routes/modules/iam-visualizer/+page.svelte
@@ -1,18 +1,11 @@
 <script lang="ts">
     import IamVisualizer from "$lib/iam-visualizer/IamVisualizer.svelte";
-    import { nodeStore } from "../../stores/nodes/nodes";
-    import { iconStore } from "../../stores/icons/icons";
-    import { page } from "$app/stores";
-    import { goto } from "$app/navigation";
+    import { nodeStore } from "../../../stores/nodes/nodes";
+    import { iconStore } from "../../../stores/icons/icons";
 
     var selectedNodeTable = nodeStore.getters.getSelectedNodeTable();
 
     async function initGraphData() {
-        // TODO: better client-side auth handling
-        if ($page.data.session === null) {
-            goto("/auth/signin");
-        }
-
         await nodeStore.actions.fetchNodes(selectedNodeTable);
         await iconStore.actions.fetchIcons();
     }
